Change the threshold type from double to float

+++ Assignment2/simulations.h
@@ -2,6 +2,6 @@
 #define simulations_h
 
 /* prototypes */
-void sensor_node(int num_rows, int num_cols, double threshold, MPI_Comm world_comm, MPI_Comm cart_comm);  // function declaration for sensor node simulation
+void sensor_node(int num_rows, int num_cols, float threshold, MPI_Comm world_comm, MPI_Comm cart_comm);  // function declaration for sensor node simulation
 
 #endif // simulations_h
